Fix plotting star imports
